+++ scripts/provision/provision.sh
@@ -8,12 +8,12 @@ echo "Sourcing scripts that set environment variables for provision scripts.."
 
 echo "Making all shell scripts executable.."
 find /vagrant/scripts/ -type f -name '*.sh' -exec chmod +x {} \;
-[[ -d /vagrant/share/misc/scripts/$(hostname -s)/ ]] && find /vagrant/share/misc/scripts/$(hostname -s)/ -type f -name '*.sh' -exec chmod +x {} \;
+[[ -d /vagrant/shared/misc/scripts/$(hostname -s)/ ]] && find /vagrant/shared/misc/scripts/$(hostname -s)/ -type f -name '*.sh' -exec chmod +x {} \;
 
 echo "Executing shell scripts.."
 
-if [[ -e "/vagrant/share/misc/scripts/$(hostname -s)/${provisioner}-pre.sh" ]]; then
-  /vagrant/share/misc/scripts/$(hostname -s)/${provisioner}-pre.sh || exit 1
+if [[ -e "/vagrant/shared/misc/scripts/$(hostname -s)/${provisioner}-pre.sh" ]]; then
+  /vagrant/shared/misc/scripts/$(hostname -s)/${provisioner}-pre.sh || exit 1
 fi
 
 if [[ -d "/vagrant/scripts/custom/${provisioner}" ]]; then
@@ -26,8 +26,8 @@ else
   done
 fi
 
-if [[ -e "/vagrant/share/misc/scripts/$(hostname -s)/${provisioner}-post.sh" ]]; then
-  /vagrant/share/misc/scripts/$(hostname -s)/${provisioner}-post.sh || exit 1
+if [[ -e "/vagrant/shared/misc/scripts/$(hostname -s)/${provisioner}-post.sh" ]]; then
+  /vagrant/shared/misc/scripts/$(hostname -s)/${provisioner}-post.sh || exit 1
 fi
 
 echo "Finishing ${0}.."